pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.117.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-mongodb>=0.2.0",
    "motor[asyncio]>=3.7.1",
    "openai-agents>=0.3.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-core>=2.33.2",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
    "pymongo==4.14.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.36.0",
]
